requirements.txt
requests>=2.31.0
praw>=7.7.1
schedule>=1.2.0
jinja2>=3.1.2
python-dotenv>=1.0.0
flask>=2.3.0
